// ui.retired/lib/game-types.ts
export interface GameView {
  type: 'character' | 'inventory' | 'quests' | 'map' | 'battle' | 'dialogue' | 'shop' | 'auction' | 'mail' | 'friends' | 'lfp' | 'profile' | 'party' | 'oghams' | 'guild' | 'bestiary' | 'achievements' | 'leaderboards' | 'skills' | 'crafting' | 'companions' | 'bank' | 'bounties' | 'mounts' | 'creatures' | 'jobs' | 'cards' | 'tournaments' | 'world_events' | 'dm'
}